codigo a compilar, a ordenar corretamente, a pesquisar corretamente, a imprimir corretamente

+++ codigo_fonte/constantes.h
@@ -5,9 +5,9 @@
 
 typedef struct{
 	char id[4];
-	short int capacidade;
+	int capacidade;
 	bool estado;
-	short int crono;
+	int crono;
 	int partem;
 	int chegam;
 } Aeroporto;
